py-api/pyproject.toml
[build-system]
requires = ["setuptools"]

[project]
dynamic=["2025.0.0"]
name="lexmpm-py-api"
dependencies = [
    "sqlalchemy >= 2.0.40",
    "sqlalchemy_utils >= 0.41.2",
    "fastapi >= 0.115.12",
    "psycopg-binary >= 3.2.7",
    "psycopg-pool >= 3.2.6",
    "pyjwt[crypto] >=2.10.1",
    "passlib[brypt] >= 1.7.4",
    "python-multipart >= 0.0.20"
]
requires-python=">=3.13"

authors=[
    {name="Robert L. Quigley", email="robert@logicaexmachina.dev"}
]
description= "The backend API for LexMPM."
readme = "README.md"
license = "MIT"
license-files= ["LICENSE"]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Topic :: ",
    "Framework :: FastAPI"
]
